scripts/requirements-e2e.txt
# E2E Validator Requirements
# Standalone - No IQE or external framework dependencies

kubernetes>=28.1.0
psycopg2-binary>=2.9.9
boto3>=1.34.0
celery>=5.3.4
redis>=5.0.1
requests>=2.31.0
click>=8.1.7
urllib3>=2.1.0
PyYAML>=6.0.1  # For nise YAML parsing
koku-nise>=2.0.0  # Test data generator

# Optional but recommended
colorama>=0.4.6  # Cross-platform colored output
tabulate>=0.9.0  # Pretty tables

